app/schemas.py
from pydantic import BaseModel, Field
from typing import Optional
from enum import Enum


class TaskStatus(str, Enum):
    TODO = "TODO"
    IN_PROGRESS = "IN_PROGRESS"
    DONE = "DONE"


class TaskBase(BaseModel):
    title: str
    status: TaskStatus = Field(default=TaskStatus.TODO)
    lead_id: str
    notes: Optional[str] = None


class TaskCreate(TaskBase):
    pass


class TaskUpdate(BaseModel):
    title: Optional[str] = None
    status: Optional[TaskStatus] = None
    notes: Optional[str] = None


class Task(TaskBase):
    id: str

    class Config:
        from_attributes = True


